requirements.txt
numpy==1.21.5
pandas==1.3.5
opencv-python==4.5.5.64
torch
torchvision
tqdm==4.62.3
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
albumentations
timm==0.8.17dev0
pytorch_lightning
open_clip_torch